+++ tests/testthat/test-class_record.R
@@ -88,7 +88,7 @@ tar_test("record_encode_field() nonempty", {
 
 tar_test("record_encode_field() with bad characters", {
   out <- record_encode_field(c("a\nb*|c\r", "b\t"))
-  expect_equal(out, "abc b ")
+  expect_equal(out, "a bc  b ")
   expect_true(nzchar(out))
   expect_length(out, 1L)
 })
@@ -96,7 +96,7 @@ tar_test("record_encode_field() with bad characters", {
 tar_test("record_encode_field() with more bad characters", {
   str <- "123abcXYZ90878734 123a*}*}*}|||*|*|*|*|bcXYZ90878734 ...,,,_-="
   out <- record_encode_field(str)
-  exp <- "123abcXYZ90878734 123abcXYZ90878734 ...,,,_"
+  exp <- "123abcXYZ90878734 123a}}}bcXYZ90878734 ...,,,_-="
   expect_equal(out, exp)
 })